+++ queries-core/src/lib.rs
@@ -141,7 +141,7 @@ impl IntentParser {
         Ok(IntentParser { classifiers: classifiers })
     }
 
-    pub fn run_intent_classifiers(&self, input: &str, probability_threshold: f64, intent_filter: Option<&[&str]>) -> Vec<IntentClassifierResult> {
+    pub fn run_intent_classifiers(&self, input: &str, probability_threshold: f32, intent_filter: Option<&[&str]>) -> Vec<IntentClassifierResult> {
         assert!(probability_threshold >= 0.0 && probability_threshold <= 1.0, "it's a developer error to pass a probability_threshold between 0.0 and 1.0");
 
         let preprocessor_result = preprocess(input);
